pyproject.toml
[project]
name = "til"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.18"
dependencies = [
    "pyyaml>=6.0.3",
]
